+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Si vous souhaitez voir le détail, cliquez sur les numéros qui vous renverrons 
 - [#422](https://github.com/Refhi/Weda-Helper/issues/422) - La déclaration de MT automatique se fait dans un onglet séparé.
 - [#422](https://github.com/Refhi/Weda-Helper/issues/422) - La fenêtre de validation de déclaration du MT se ferme automatiquement.
 - [#408](https://github.com/Refhi/Weda-Helper/issues/408) - La fenêtre de prévisualisation de l'arrêt de travail se ferme automatiquement.
+- [#409](https://github.com/Refhi/Weda-Helper/issues/409) - On peut désormais éditer un atcd directement depuis la page d'accueil du dossier patient.
 
 ## Bugfix :
 - [#415](https://github.com/Refhi/Weda-Helper/issues/415) - La clé API n'était plus générée automatiquement depuis quelques versions => de nouveau générée automatiquement.